backend/requirements.txt
flask
flask_cors
kafka-python
